+++ local_examples/test_applications.R
@@ -1,0 +1,42 @@
+#
+# ----- App #1:  Successful GET ------------------------------------------------
+
+library(beakr)
+
+newBeakr() %>%
+  get("/", function(req, res, err) {
+    return("Successful GET request!\n")
+  }) %>%
+  startBeakr()
+
+# In browser:
+#   http://127.0.0.1:8080
+# > Successful GET request!
+
+# ----- App #2:  Successful POST -----------------------------------------------
+
+library(beakr)
+
+newBeakr() %>%
+  post("/", function(req, res, err) {
+    return("Successful POST request!\n")
+  }) %>%
+  startBeakr()
+
+# In terminal:
+#  curl -X POST http://127.0.0.1:8080/
+# > Successful POST request!
+
+# ----- App #3:  Successful PUT ------------------------------------------------
+
+library(beakr)
+
+newBeakr() %>%
+  put("/", function(req, res, err) {
+    return("Successful PUT request!\n")
+  }) %>%
+  startBeakr()
+
+# In terminal:
+#  curl -X PUT http://127.0.0.1:8080/
+# > Successful PUT request!